# doc/swagger/backend.json
{"openapi":"3.0.0","paths":{"/bff/v1/healthcheck":{"get":{"operationId":"healthcheck","summary":"Healthcheck","description":"Chequeo de Estado Api","parameters":[],"responses":{"200":{"description":""}},"tags":["health"]}},"/bff/v1/auth-users/register":{"post":{"operationId":"AuthUsersController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRegisterUserDto"}}}},"responses":{"201":{"description":""}}}},"/bff/v1/auth-users/confirmation-register":{"post":{"operationId":"AuthUsersController_confirmation","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmRegisterDto"}}}},"responses":{"201":{"description":""}}}},"/bff/v1/auth-users/resend-confirmation":{"post":{"operationId":"AuthUsersController_resendConfirmation","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResendConfirmRegisterDto"}}}},"responses":{"201":{"description":""}}}},"/bff/v1/auth-users/send-code":{"post":{"operationId":"AuthUsersController_forgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResendConfirmRegisterDto"}}}},"responses":{"201":{"description":""}}}},"/bff/v1/auth-users/validation-code-password":{"post":{"operationId":"AuthUsersController_validationCodeAndPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"201":{"description":""}}}},"/bff/v1/auth-users/login":{"post":{"operationId":"AuthUsersController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthLoginUserDto"}}}},"responses":{"201":{"description":""}}}},"/bff/v1/auth-users/sign-out":{"post":{"operationId":"AuthUsersController_signOut","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResendConfirmRegisterDto"}}}},"responses":{"201":{"description":""}}}},"/bff/v1/auth-users/group-projects":{"post":{"operationId":"AuthUsersController_saveGroupProjects","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupProjectsDto"}}}},"responses":{"201":{"description":""}}}},"/bff/v1/auth-users/{id}":{"get":{"operationId":"AuthUsersController_readGroupProjects","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"AuthUsersController_updateGroupProjects","parameters":[],"responses":{"200":{"description":""}}},"delete":{"operationId":"AuthUsersController_deleteGroupProjects","parameters":[],"responses":{"200":{"description":""}}}},"/bff/v1/auth-users/all-group-projects":{"post":{"operationId":"AuthUsersController_allGroupProjects","parameters":[],"responses":{"201":{"description":""}}}},"/bff/v1/auth-users/profile-create":{"post":{"operationId":"AuthUsersController_profileCreate","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfilesDto"}}}},"responses":{"201":{"description":""}}}},"/bff/v1/auth-users/profile-all":{"post":{"operationId":"AuthUsersController_getProfiles","parameters":[],"responses":{"201":{"description":""}}}},"/bff/v1/bff-gateway":{"post":{"operationId":"createForm","summary":"creacion de formulario","description":"Permite registro de formulario y envio de Emails (Agradecimiento y Cuestionario de preguntas)","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FormRequest"}}}},"responses":{"201":{"description":""}},"tags":["Formulario contacto"]}},"/bff/v1/bff-gateway/validateForm":{"post":{"operationId":"GatewayController_validateForm","summary":"validacion formulario","description":"Permite validacion formulario (Top Limit Form))","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationForm"}}}},"responses":{"201":{"description":""}},"tags":["Formulario contacto"]}},"/bff/v1/bff-gateway/survey":{"post":{"operationId":"GatewayController_saveSurvey","summary":"save CES","description":"Permite registros CES","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSatisfactionSurveyDto"}}}},"responses":{"201":{"description":""}},"tags":["Formulario contacto"]}},"/bff/v1/feedback":{"post":{"operationId":"saveFeedback","summary":"save feedbak","description":"Permite el registro de feedbak","parameters":[],"responses":{"201":{"description":""}},"tags":["Feedback"]},"get":{"operationId":"findAllFeedback","summary":"findAll feedbak","description":"Permite la consulta de feedbak","parameters":[],"responses":{"200":{"description":""}},"tags":["Feedback"]}},"/bff/v1/generators-token":{"get":{"operationId":"getGenerateToken","summary":"get token jwt","description":"Permite la generacion de token aleatorio","parameters":[],"responses":{"200":{"description":""}},"tags":["GeneratorsToken"]}},"/bff/v1/bulk-data-upload":{"post":{"operationId":"saveFileUpload","summary":"upload file excel","description":"Permite subir un archivo excel, para su posterior guardado en BD","parameters":[],"responses":{"201":{"description":""}}},"get":{"operationId":"saveFileUpload","summary":"realiza una busquedad de precio x producto","description":"Prealiza una busquedad de precio x producto","parameters":[{"name":"slug","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/bff/v1/bulk-data-upload/{id}":{"delete":{"operationId":"saveFileUpload","summary":"eliminacion de informacion de producto","description":"eliminacion de informacion de producto","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"updateProductPrice","summary":"actualizacion de informacion producto","description":"eliminacion de informacion producto","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBulkDataUploadDto"}}}},"responses":{"200":{"description":""}}}}},"info":{"title":"Bff contact API","description":"Hermes contact Bff","version":"0.2.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:5500","description":"Servidor local"}],"components":{"schemas":{"AuthRegisterUserDto":{"type":"object","properties":{"name":{"type":"string","description":"Nombre Usuario"},"email":{"type":"string","description":"Email usuario"},"access_general":{"type":"string","description":"Acceso General"},"privileges":{"description":"Privileges","type":"array","items":{"type":"string"}},"password":{"type":"string","description":"clave usuario"}},"required":["name","email","access_general","privileges","password"]},"ConfirmRegisterDto":{"type":"object","properties":{"username":{"type":"string","description":"Nombre Usuario"},"code":{"type":"string","description":"code confirmacion registro Usuario"}},"required":["username","code"]},"ResendConfirmRegisterDto":{"type":"object","properties":{"username":{"type":"string","description":"Nombre Usuario"}},"required":["username"]},"ChangePasswordDto":{"type":"object","properties":{"username":{"type":"string","description":"Nombre Usuario"},"verificationCode":{"type":"string","description":"code confirmacion registro Usuario"},"password":{"type":"string","description":"code confirmacion registro Usuario"}},"required":["username","verificationCode","password"]},"AuthLoginUserDto":{"type":"object","properties":{"email":{"type":"string","description":"Email usuario"},"password":{"type":"string","description":"clave usuario"}},"required":["email","password"]},"GroupProjectsDto":{"type":"object","properties":{"name":{"type":"string","description":"Nombre Grupo"},"projects":{"description":"Projectos","type":"array","items":{"type":"string"}}},"required":["name","projects"]},"ProfilesDto":{"type":"object","properties":{"name":{"type":"string","description":"Profiles"}},"required":["name"]},"FormField":{"type":"object","properties":{"reportIdentification":{"type":"string","description":"Identificacion de reporte"},"fieldsContainer":{"description":"fieldsContainer","type":"array","items":{"type":"string"}}},"required":["reportIdentification","fieldsContainer"]},"FormRequest":{"type":"object","properties":{"contactName":{"type":"string","description":"Nombre del contacto","example":"Juan"},"contactLastName":{"type":"string","description":"Apellido del contacto","example":"Perez"},"contactEmail":{"type":"string","description":"Email del contacto","example":"info@gmail.com"},"projectSlug":{"type":"string","description":"projecto","example":"PORI"},"fields":{"description":"Campo fields","allOf":[{"$ref":"#/components/schemas/FormField"}]},"blobs":{"description":"Archivos adjuntos","type":"array","items":{"type":"string"}}},"required":["contactName","contactLastName","contactEmail","projectSlug","fields"]},"ValidationForm":{"type":"object","properties":{"slug":{"type":"string","description":"slug"}},"required":["slug"]},"CreateSatisfactionSurveyDto":{"type":"object","properties":{"level":{"type":"number","description":"nivel de satisfaccion"},"name":{"type":"string","description":"nombre contacto form"},"email":{"type":"string","description":"email del contacto"},"project":{"type":"string","description":"name Projecto"}},"required":["level","name","email","project"]},"UpdateBulkDataUploadDto":{"type":"object","properties":{"product_slug":{"type":"string","description":"product_slug"},"price":{"type":"string","description":"price"},"currency":{"type":"string","description":"currency"},"pill":{"type":"string","description":"pill"},"price_description":{"type":"string","description":"price_description"},"iva":{"type":"string","description":"price_description"},"discount":{"type":"string","description":"discount"}}}}},"externalDocs":{"url":"http://localhost:5500/contact-bff/docs-json","description":"Documentación BFF formato json"}}